feat: 이름 유효성 검사 추가
