+++ app/src/main/res/values/strings.xml
@@ -6,4 +6,6 @@
     <string name="sign_up_password">Password</string>
     <string name="sign_up_password_confirm">PasswordConfirm</string>
     <string name="sign_up_button">Sign Up</string>
+    <string name="sign_up_username_input_validation_length_message">이름은 2~5자여야 합니다.</string>
+    <string name="sign_up_username_input_validation_message">이름에는 숫자나 기호가 포함될 수 없습니다.</string>
 </resources>
